Adjusts stationSignalReconstructor to changed event structure and renames it to electricFieldSignalReconstructor
